backend/requirements.txt
# Sydney Guide Backend Requirements
# FastMCP ve API bagimlilikları

# Official MCP Python SDK - FastAPI Integration
mcp[server]>=1.0.0
fastmcp>=0.9.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Clients
httpx>=0.25.0
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Location Services
geopy>=2.4.0
googlemaps>=4.10.0

# Logging
structlog>=23.2.0

# Environment Variables
python-dotenv>=1.0.0

# Type Hints
typing-extensions>=4.8.0

# WebSocket Support
websockets>=12.0

# JSON Processing
pydantic>=2.5.0

# Date/Time
python-dateutil>=2.8.0

# Async Task Queue (optional)
celery>=5.3.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.0.0 